fix: scripts: lintChanged.sh: unshallow branch
Unshallow the current working branch or HEAD if it is detected.
This addresses a problem when we shallow fetch (depth=1)
origin/main, i.e. merge main into a PR branch. This would
lead the `git merge-base` to fail as we wouldn't have
enough history of our PR branch to run.